start work on spectram correlation routines
As an aid to identifying likely library components to select for
performing a decomposition, the cross-correlation between the residual
spectrum and the component libraries will be shown.  A correlation
value near 1.0 indicates a high probability that this library component
is present in the residual spectrum and should be included in the
decomposition.  But, false positives can occur, so only the 
decomposition results can show whether or not a library commponent
is actually present.