fix controls display incorrect results when created
The classes *Holder have a static reference to an object needed 
to update the display controls.  When the control is created it 
subscribes to future update events.  But until an update event
occurs, the display shows incorrect results.  

Now, when the control is created, it gets a reference to the 
latest object and then displays correct results.  Future updates
to the object trigger events and updates as usual.  

The static reference in the *Holder class is not used again.