add ability to calculate and apply signal zero offset correction
sometimes a mass spec ion detector shows a consistent offset 
from zero for all measurements.  This offset can be positive 
or negative and is typically low in absolute value.  It introduces 
error into measurements and can be signifcant for measurements 
of low abundance peaks.  

To correct for this detector offset, the spectrum is examined for 
signal measurements at m/z 5, 6, and 7 where the ion abundance should 
be zero.  The middle value is assumed to be the detector offset.  

A zero offset correction is applied by subtracting the detector 
offset from the signal value for each ion in the spectrum.  Various 
methods allow offset correction to be calculated, applied, or 
removed (reset).