FOODAUTH-48 Expand enums and implement the (full) working algorithm
This is the (almost) complete collection of all the parts of the
ICOSHIFT algorithm that are needed to align multiple selected
spectra of one dataset.

All necessary options can be accessed by selecting the respective
enumeration. This way a variety of combinations is possible regarding
how to align the dataset:
* The target calculation (this can optionally be an actual signal or
  a synthetic spectrum) can be set to "MEAN", "MEDIAN", or "MAX".
* The intervals to be aligned can be set to "SINGLE_PEAK",
  "WHOLE_SPECTRUM", "NUMBER_OF_INTERVALS", "INTERVAL_LENGTH", or
  "USER_DEFINED_INTERVALS". The selected intervals will be aligned to
  the previously calculated target.
* The type of shift correction can be set to "FAST", "BEST", or
  "USER_DEFINED". This will affect how the individual shifts will be
  calculated by changing the observed part of data.
* The gap filling type can be set to "ZERO" or "MARGIN" and defines
  how "missing" data points will be handled after the alignment.

Starting with a directory which contains multiple spectra, the returned
result is an object of the type SimpleMatrix which contains the aligned
dataset.