Expand the zero filling for processed data
To cope with the fact that the size of the processed data after import
does not correspond to its original size, the zero filling was extended
accordingly. Now the correct size will be restored.
In addition, the descriptions in the enum were tidied up and the number
of data points was corrected.