FOODAUTH-41 Implement the requested PR changes and add cleaner code
The security checks regarding the calculation of the calibration target
have been removed because they were useless. Instead one check was
introduced in the calculateCalibrationTarget method to let the user
know that the value of the scale of the CauchyDistribution must be
greater than zero.
In addition, the code where methods from EJML are used, and it's
applicable, has been re-factored to make use of the built in
functionality to access data instead of using simple loops.