Fix the determination and removal of digital filter
The fix reverts the determination to the old structure where GRPDLY not
only had to be present but it has to have a reasonable value. Beside
that the overall determination is not finished if the GRPDLY value was
checked, but also DECIM and DSPFVS were considered for an alternative
determination of the necessary value. This gives a complete test.