NMR-78 Fix the orientation of displayed Bruker nmr spectra
Until now the spectral data was added to the signals in the necessary
orientation for diplaying purposes (e.g. from left to right) while the
chemical shift axis was generated the other way round (e.g. from right
to left) and stored alongside the nmr intensities. This fix will
reorganize the imported and processed Bruker nmr spectra in relation to
their respective chemical shift axis.
This seems to affect only Bruker data because handling e.g. JCAMP-DX
data this way will reverse the spectral data unnecessary. Therefore the
FourierTransformationProcessor is adopted to check for filtered Bruker
data and change the direction of spectral data. Otherwise the signals
will be added without further modifications.
Now that the sorting/direction of fid signals and nmr signals is written
down in the API (fid: 0 to AcquisitionTime; spectrum: from biggest
number to 0 and beyond) it is necessary to make sure every data type is
saved correct in the object model so that it can be drawed and used in
the right way by other plugins.