FOODAUTH-264 Implement proper checking/comparing of signals size
An additional method was added to check if the length of data is of
power of two because just comparing two data sets that meets the
requirements of the processor but have different length will always
let the FT processor fail. The different lengths originate from the
fixed selection ZeroFillingFactor.AUTO that will always fill up to
the next power of two!