Fixed #526 - Plugin Installer - 'Don't ask again.' flag is not tracked
The 'Don't ask again.' flag was not tracked and dialog was always shown.
This was a bit tricky to detect as it was hidden by the previous feature
check. The Chromatotec converter was not listed before. It got obvious
when only installing the Chromatotec converter. Then the dialog
preference option was retrieved, but the dialog was displayed regardless
of the selected don't ask again flag.